+++ src/php/Compiler/Domain/Analyzer/Environment/GlobalEnvironment.php
@@ -40,7 +40,7 @@ final class GlobalEnvironment implements GlobalEnvironmentInterface
     /** @var array<string, array<string, Symbol>> */
     private array $interfaces = [];
 
-    private bool $allowPrivateAccess = false;
+    private int $allowPrivateAccessCounter = 0;
 
     public function __construct()
     {
@@ -179,9 +179,14 @@ final class GlobalEnvironment implements GlobalEnvironmentInterface
         return null;
     }
 
-    public function setAllowPrivateAccess(bool $allowPrivateAccess): void
+    public function addLevelToAllowPrivateAccess(): void
     {
-        $this->allowPrivateAccess = $allowPrivateAccess;
+        ++$this->allowPrivateAccessCounter;
+    }
+
+    public function removeLevelToAllowPrivateAccess(): void
+    {
+        --$this->allowPrivateAccessCounter;
     }
 
     public function addInterface(string $namespace, Symbol $name): void
@@ -309,6 +314,7 @@ final class GlobalEnvironment implements GlobalEnvironmentInterface
 
     private function isPrivateDefinitionAllowed(PersistentMapInterface $meta): bool
     {
-        return $this->allowPrivateAccess || !$meta[Keyword::create('private')];
+        return $this->allowPrivateAccessCounter > 0
+            || !$meta[Keyword::create('private')];
     }
 }